add rewardRate to the nodeStakingUdpate txn synthetic record

+++ hedera-node/src/main/java/com/hedera/services/context/properties/GlobalDynamicProperties.java
@@ -122,7 +122,7 @@ public class GlobalDynamicProperties {
 	private KnownBlockValues knownBlockValues;
 	private int maxReturnedNftsPerTouch;
 	private Gas exchangeRateGasReq;
-	private double stakingRewardRate;
+	private long stakingRewardRate;
 	private long stakingStartThreshold;
 	private int nodeRewardPercent;
 	private int stakingRewardPercent;
@@ -225,7 +225,7 @@ public class GlobalDynamicProperties {
 		maxReturnedNftsPerTouch = properties.getIntProperty("autoRemove.maxReturnedNftsPerTouch");
 		knownBlockValues = properties.getBlockValuesProperty("contracts.knownBlockHash");
 		exchangeRateGasReq = Gas.of(properties.getLongProperty("contracts.precompile.exchangeRateGasCost"));
-		stakingRewardRate = properties.getDoubleProperty("staking.rewardRate");
+		stakingRewardRate = properties.getLongProperty("staking.rewardRate");
 		stakingStartThreshold = properties.getLongProperty("staking.startThreshold");
 		nodeRewardPercent = properties.getIntProperty("staking.fees.nodeRewardPercentage");
 		stakingRewardPercent = properties.getIntProperty("staking.fees.stakingRewardPercentage");
@@ -548,7 +548,7 @@ public class GlobalDynamicProperties {
 		return exchangeRateGasReq;
 	}
 
-	public double getStakingRewardRate() {
+	public long getStakingRewardRate() {
 		return stakingRewardRate;
 	}
